+++ app/src/main/java/model/GetMyPublishResponse.java
@@ -23,7 +23,7 @@ public class GetMyPublishResponse {
         private String commodityId;
         private String commodityName;
         private String commodityDescription;
-        private int commodityValue;
+        private Double commodityValue;
         private String commodityClass;
         private String commodityImage;
         private String commodityStatus;
