+++ app/src/main/java/model/GetMySellResponse.java
@@ -25,7 +25,7 @@ public class GetMySellResponse {
         private String commodityId;
         private String commodityName;
         private String commodityDescription;
-        private int commodityValue;
+        private Double commodityValue;
         private String commodityClass;
         private String commodityImage;
         private String buyerId;
